Implement test (failed).
The test fails **because** I'm using a singleton for my
`TrackingService` instance. At the end of the first test
`TrackingServiceColumnFixture`, I had added 1 to the total. Because of
this history, I start my test with a total of 1. So after adding 10 five
times, the test will fail because the total is 51 **not** 50.

I learned of this error from the video, but how could I discover it
myself?